backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-extra-types==2.1.0
python-dotenv==1.0.0
supabase==2.0.3
httpx==0.24.1
python-jose==3.3.0
python-jwt==4.0.0
email-validator==2.0.0
typing-extensions==4.8.0 